src/requirements.txt
Pillow==10.0.1
numpy==1.26.1
requests==2.31.0
matplotlib==3.8.2
click==8.1.7